skills/.experimental/product-init/scripts/requirements.txt
pyyaml
jsonschema
python-frontmatter
requests
